+++ src/Steps/Refiners/RefinerInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Crwlr\Crawler\Steps\Refiners;
+
+use Psr\Log\LoggerInterface;
+
+interface RefinerInterface
+{
+    public function refine(mixed $value): mixed;
+
+    public function addLogger(LoggerInterface $logger): static;
+}